New heuristics

+++ src/triple_bwt.hpp
@@ -786,7 +786,7 @@ private:
 
         RpqTree rpqTree(rpq, predicates_map, real_max_P);
         auto mandData = rpqTree.getMandatoryData();
-        selectivity::info sel_min{std::numeric_limits<double>::max(), selectivity::source};
+        selectivity::info sel_min{std::numeric_limits<uint64_t>::max(), selectivity::source};
         uint64_t i_split = 0;
         uint64_t sigma = (max_O > max_S) ? max_O : max_S;
         const auto& pos_pred_vec = mandData.pos_pred;
@@ -862,7 +862,7 @@ private:
                                                      unordered_map<std::string, uint64_t> &predicates_map){
 
 
-        selectivity::info sel_min{std::numeric_limits<double>::max(), selectivity::source};
+        selectivity::info sel_min{std::numeric_limits<uint64_t>::max(), selectivity::source};
         uint64_t i_split = 0;
         uint64_t sigma = (max_O > max_S) ? max_O : max_S;
         const auto& pos_pred_vec = mandData.pos_pred;
